+++ main.go
@@ -8,7 +8,7 @@ import (
 )
 
 const (
-	version = "0.2.0"
+	version = "0.2.1"
 )
 
 func main() {