+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 * Clean up acceptance test logics by moving to config struct
 * Back up ACLs
 * Back up Prepared Queries
+* Optionally encrypt data with a CRYPTO_PASSWORD environment variable passphrase
+* Add meta on backups
+* Refactor backup layout and adjust restores
 
 ## 0.1.5
 * Fix regression in temp file restore location